+++ CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- nill
-
+### Added
+- `<OmniSearch />` search bar combined with `useAlgolia()`
 
 ## [0.1.3] - 2020-06-17
 ### Changed